fix: Move character utility functions to separate utils file for better testability
